daemon: extract resolution module from config.rs (#41 cont)
Pulls all of the `resolve_*` config helpers out of `config/mod.rs`
into `config/resolution.rs`:

- `resolve_llm_config`, `resolve_backend_tasks_llm_config`,
  `resolve_purpose_llm_config`, `resolve_connection_llm_config`,
  `resolve_llm_config_from`
- `resolve_embeddings_config`, `resolve_persistence_config`,
  `resolve_database_config`
- `resolve_context_budget`, `purpose_max_context_override`,
  `DEFAULT_PURPOSE_MAX_CONTEXT_TOKENS`
- Connector-default helpers (`parse_connector_or_openai`,
  `default_base_url`, `default_llm_model`,
  `default_backend_llm_model`, `normalize_optional_value`,
  `default_embedding_model`)

Env-var helpers in `mod.rs` (`default_api_key_env` /
`default_model_env` / `default_base_url_env`,
`default_database_max_connections`, `default_push_on_update`)
are now `pub(super)` so resolution.rs can call them.

mod.rs is now 2453 lines (from 2912 before this change, 4373 at
the start of #41). Public API at `config::*` is unchanged.
